Rewrote the math for leg positioning.
This is the first major change to pythonifying the quadrup3d code. The
math used to change the position of the legs has been redone and the
coordinate system for choosing said position has been made universal
across all legs. This makes it significantly easier to for instance
translate the body in the X and Y axis.